Added Pages to the StatusBoard. Now the carousel displays several Pages such as General and Pull Requests, each made of components.
