+++ app/assets/stylesheets/static_pages.scss
@@ -42,12 +42,16 @@ marquee {
   padding-top: $small-pad;
 }
 
+#page-title {
+  text-align: center;
+}
+
 #carousel {
-  height: 1000px;
+  height: 90vh;
 }
 
 #twitter {
-  width: 20%;
+  width: 30%;
   padding-left: 35px;
   padding-top: 40px;
   float: left;
@@ -135,8 +139,9 @@ marquee {
 
 #git {
   float: left;
-  width: 40%;
-  padding-left: 5px;
+  width: 50%;
+  padding-left: 25px;
+  padding-right: 25px;
   padding-top: 40px;
 
   #git-feed {
