--- app/models/components/twitter_component.rb
@@ -1,7 +1,0 @@
-module Components
-  class TwitterComponent
-    def tweets
-      @tweets ||= Tweet.latest(3)
-    end
-  end
-end
