+++ app/models/pages/components/twitter_component.rb
@@ -1,0 +1,9 @@
+module Pages
+  module Components
+    class TwitterComponent
+      def tweets
+        @tweets ||= Tweet.latest(3)
+      end
+    end
+  end
+end
