+++ app/models/pages/header_page.rb
@@ -1,0 +1,11 @@
+module Pages
+  class HeaderPage < StatusBoardPage
+    def github_component
+      @github_component ||= Components::GithubComponent.new
+    end
+
+    def weather_component
+      @weather_component ||= Components::WeatherComponent.new
+    end
+  end
+end
