+++ app/models/pages/pull_requests_page.rb
@@ -1,0 +1,7 @@
+module Pages
+  class PullRequestsPage < StatusBoardPage
+    def github_component
+      @github_component ||= Components::GithubComponent.new
+    end
+  end
+end
