+++ app/models/pages/status_board_page.rb
@@ -1,0 +1,4 @@
+module Pages
+  class StatusBoardPage
+  end
+end
