--- app/models/presenters/components/photo_component.rb
@@ -1,11 +1,0 @@
-module Presenters
-  module Components
-    class PhotoComponent < StatusBoardComponent
-
-      #this will need to be updated to check if there's actually an image
-      def has_image_url?
-        !url.empty?
-      end
-    end
-  end
-end
