--- app/models/presenters/components/status_board_component.rb
@@ -1,9 +1,0 @@
-module Presenters
-  module Components
-    class StatusBoardComponent < Base
-      def partial
-        @wrapped.class.name.underscore
-      end
-    end
-  end
-end
