--- app/models/presenters/components/weather_component.rb
@@ -1,6 +1,0 @@
-module Presenters
-  module Components
-    class WeatherComponent < StatusBoardComponent
-    end
-  end
-end
