+++ app/models/presenters/pages/components/github_component.rb
@@ -1,0 +1,8 @@
+module Presenters
+  module Pages
+    module Components
+      class GithubComponent < Base
+      end
+    end
+  end
+end
