+++ app/models/presenters/pages/components/weather_component.rb
@@ -1,0 +1,8 @@
+module Presenters
+  module Pages
+    module Components
+      class WeatherComponent < Base
+      end
+    end
+  end
+end
