+++ app/models/presenters/pages/general_page.rb
@@ -1,0 +1,7 @@
+module Presenters
+  module Pages
+    class GeneralPage < PagePresenter
+      presents :twitter_component, :weather_component
+    end
+  end
+end
