+++ app/models/presenters/pages/header_page.rb
@@ -1,0 +1,7 @@
+module Presenters
+  module Pages
+    class HeaderPage < PagePresenter
+      presents :github_component, :weather_component
+    end
+  end
+end
