+++ app/models/presenters/pages/pull_requests_page.rb
@@ -1,0 +1,7 @@
+module Presenters
+  module Pages
+    class PullRequestsPage < PagePresenter
+      presents :github_component
+    end
+  end
+end
