+++ app/views/components/_photo.html.erb
