+++ app/views/components/_weather.html.erb
