+++ app/views/pages/_general_page.html.erb
@@ -1,0 +1,5 @@
+<h1 id="page-title">General</h1>
+
+<%= render '/components/twitter', component: page.twitter_component %>
+
+<%= render '/components/weather', component: page.weather_component %>
