+++ app/views/pages/_pull_requests_page.html.erb
@@ -1,0 +1,3 @@
+<h1 id="page-title">Pull Requests</h1>
+
+<%= render '/components/pull_requests', component: page.github_component %>
