+++ app/views/static_pages/_indicators.html.erb
@@ -1,7 +1,6 @@
 <!-- Indicators -->
 <ol class="carousel-indicators">
-  <li data-target="#carousel" data-slide-to="0" class="active"></li>
-  <% (1..@status_board.components.length-1).each do |i| %>
-    <li data-target="#carousel" data-slide-to="<%= i %>"></li>
+  <% @status_board.pages.length.times do |i| %>
+    <li data-target="#carousel" data-slide-to="<%= i %>"<% if i.zero? %> class="active"<% end %>></li>
   <% end %>
 </ol>
